Add shutdown() to Dispatcher interface
I wire up Dispatcher through Spring Configuration annotations, and want to
give Dispatcher a chance to shutdown gracefully when Spring is shutting
down. Since shutdown() isn't part of the Dispatcher interface, I get
warnings about a method that doesn't exist.

This adds shutdown() to the interface to fix that warning.